refactoring :: CompaniesDomainInterface Transfer 나눔
