+++ Projects/Domain/CompaniesDomain/Interface/DTO/Response/CompanyResponseDTO.swift
@@ -1,0 +1,22 @@
+import Foundation
+
+public struct CompanyResponseDTO: Decodable {
+    public let id: Int
+    public let name: String
+    public let logoURL: String
+    public let take: Int
+
+    public init(id: Int, name: String, logoURL: String, take: Int) {
+        self.id = id
+        self.name = name
+        self.logoURL = logoURL
+        self.take = take
+    }
+
+    enum CodingKeys: String, CodingKey {
+        case id
+        case name
+        case logoURL = "logo_url"
+        case take
+    }
+}
