+++ Projects/Domain/CompaniesDomain/Interface/Entity/StudentCompanyListEntity.swift
@@ -7,17 +7,3 @@ public struct StudentCompanyListEntity: Equatable {
         self.companies = companies
     }
 }
-
-public struct CompanyEntity: Equatable, Hashable {
-    public let id: Int
-    public let name: String
-    public let logoURL: String
-    public let take: Int
-
-    public init(id: Int, name: String, logoURL: String, take: Int) {
-        self.id = id
-        self.name = name
-        self.logoURL = logoURL
-        self.take = take
-    }
-}
